requirements.txt
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.2
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
coverage==7.3.4
cryptography==41.0.7
dill==0.3.7
ecdsa==0.18.0
fastapi==0.105.0
flake8==6.1.0
flake8-for-pycharm==0.4.1
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.2
passlib==1.7.4
platformdirs==4.1.0
pluggy==1.3.0
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==3.1.0
pylint==3.0.3
pytest==7.4.3
pytest-cov==4.1.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomlkit==0.12.3
typing_extensions==4.9.0
uvicorn==0.24.0.post1
validators==0.22.0
